apollo_dashboard: update consensus_block_number_progress_is_slow threshold and module (#8431)
