+++ crates/apollo_dashboard/resources/dev_grafana_alerts_mainnet.json
@@ -684,7 +684,7 @@
         {
           "evaluator": {
             "params": [
-              25.0
+              10.0
             ],
             "type": "lt"
           },
