+++ crates/apollo_dashboard/resources/dev_grafana_alerts_testnet.json
@@ -711,7 +711,7 @@
         {
           "evaluator": {
             "params": [
-              25.0
+              10.0
             ],
             "type": "lt"
           },
