+++ crates/apollo_dashboard/src/alert_scenarios/block_production_delay.rs
@@ -1,4 +1,4 @@
-use apollo_consensus::metrics::CONSENSUS_ROUND_ABOVE_ZERO;
+use apollo_consensus::metrics::{CONSENSUS_BLOCK_NUMBER, CONSENSUS_ROUND_ABOVE_ZERO};
 use apollo_consensus_manager::metrics::CONSENSUS_NUM_CONNECTED_PEERS;
 use apollo_consensus_orchestrator::metrics::CENDE_WRITE_BLOB_FAILURE;
 
@@ -167,4 +167,42 @@ pub(crate) fn get_cende_write_blob_failure_once_alert() -> Alert {
         AlertSeverity::Informational,
         AlertEnvFiltering::All,
     )
+}
+
+/// Block number progressed slowly (< 10) in the last 2 minutes.
+fn get_consensus_block_number_progress_is_slow(
+    alert_env_filtering: AlertEnvFiltering,
+    alert_severity: AlertSeverity,
+) -> Alert {
+    Alert::new(
+        "get_consensus_block_number_progress_is_slow",
+        "Consensus block number progress is slow",
+        AlertGroup::Consensus,
+        format!(
+            "sum(increase({}[2m])) or vector(0)",
+            CONSENSUS_BLOCK_NUMBER.get_name_with_filter()
+        ),
+        vec![AlertCondition {
+            comparison_op: AlertComparisonOp::LessThan,
+            comparison_value: 10.0,
+            logical_op: AlertLogicalOp::And,
+        }],
+        PENDING_DURATION_DEFAULT,
+        EVALUATION_INTERVAL_SEC_DEFAULT,
+        alert_severity,
+        alert_env_filtering,
+    )
+}
+
+pub(crate) fn get_consensus_block_number_progress_is_slow_vec() -> Vec<Alert> {
+    vec![
+        get_consensus_block_number_progress_is_slow(
+            AlertEnvFiltering::MainnetStyleAlerts,
+            AlertSeverity::Sos,
+        ),
+        get_consensus_block_number_progress_is_slow(
+            AlertEnvFiltering::TestnetStyleAlerts,
+            AlertSeverity::WorkingHours,
+        ),
+    ]
 }
